Cleaned up to_string functions
